requirements.txt

django-rest-framework
